templates/python/langchain/requirements.txt
parliament-functions==0.1.0
langchain
openai
